feat: recruit-org12 공고 _ 그룹 생성
- dto에서 private 없는 곳은 추가해둠

+++ backend/ttabong/src/main/java/com/ttabong/dto/recruit/requestDto/org/CreateRecruitRequestDto.java
@@ -5,6 +5,7 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.Date;
@@ -14,12 +15,11 @@ import java.util.Date;
 @AllArgsConstructor
 @Builder
 public class CreateRecruitRequestDto {
-
     private Integer templateId; // 템플릿 ID
     private LocalDateTime deadline; // 지원 마감일
     private Date activityDate; // 활동 날짜
-    private double activityStart; // 활동 시작 시간 (14.00 → 14:00)
-    private double activityEnd; // 활동 종료 시간 (20.00 → 20:00)
+    private BigDecimal activityStart; // 활동 시작 시간 (14.00 → 14:00)
+    private BigDecimal activityEnd; // 활동 종료 시간 (20.00 → 20:00)
     private Integer maxVolunteer; // 최대 모집 인원
 
 }
